slit navigation into multiple files
